+++ src/components/DSBP/index.scss
@@ -87,3 +87,8 @@
   white-space: nowrap;
 }
 }
+a.tabView {
+  cursor: pointer;
+  color: #003DA5;
+  text-decoration: none;
+}
